In trustregions, now, REJ always implies TR- :
We have a rejection iff:
   model_decreased && rho > options.rho_prime
is false.

That is, either
model_decreased is false, or
rho is nan, or
rho <= rho_prime, or
multiple of these things happen.

Note that rho_prime < 1/4, by assert(...).

We have a TR- iff:
   rho < 1/4 || ~model_decreased || isnan(rho)
is true.

If model_decreased is false, TR-.
If rho is nan, TR-.
If rho <= rho_prime < 1/4, TR-.

So it is guaranteed, indeed.