Added (optional) field veclen: Useful for productmanifold
productmanifold requires to know the length of the vec output
for each element manifold. Currently, this length is computed
by measuring the length of a zero vector at a random point.
This however requires the generation of a random point in the manifold,
which can in some cases produce an unnecessary overhead.

The new (optional) field veclen should provide the size of the output
for vec in a particular manifold.
For now, this has been set for euclideanfactory and stiefelstackedfactory.